+++ src/EntityGraphQL/Directives/IncludeDirectiveProcessor.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq.Expressions;
 using EntityGraphQL.Compiler;
@@ -9,6 +10,9 @@ namespace EntityGraphQL.Directives
     {
         public override string Name { get => "include"; }
         public override string Description { get => "Directs the executor to include this field or fragment only when the `if` argument is true."; }
+
+        public override List<ExecutableDirectiveLocation> On => new() { ExecutableDirectiveLocation.FIELD, ExecutableDirectiveLocation.FRAGMENT_SPREAD, ExecutableDirectiveLocation.INLINE_FRAGMENT };
+
         public override Expression? ProcessExpression(Expression expression, object? arguments)
         {
             if (arguments is null)