+++ src/EntityGraphQL/Directives/SkipDirectiveProcessor.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq.Expressions;
 using EntityGraphQL.Compiler;
@@ -9,6 +10,7 @@ namespace EntityGraphQL.Directives
     {
         public override string Name { get => "skip"; }
         public override string Description { get => "Directs the executor to skip this field or fragment when the `if` argument is true."; }
+        public override List<ExecutableDirectiveLocation> On => new() { ExecutableDirectiveLocation.FIELD, ExecutableDirectiveLocation.FRAGMENT_SPREAD, ExecutableDirectiveLocation.INLINE_FRAGMENT };
         public override Expression? ProcessExpression(Expression expression, object? arguments)
         {
             if (arguments is null)