+++ src/EntityGraphQL/Schema/SchemaBuilder.cs
@@ -14,7 +14,7 @@ using EntityGraphQL.Schema.FieldExtensions;
 namespace EntityGraphQL.Schema
 {
     /// <summary>
-    /// A simple schema provider to automattically create a query schema based on an object.
+    /// A simple schema provider to automatically create a query schema based on an object.
     /// Commonly used with a DbContext.
     /// </summary>
     public static class SchemaBuilder
@@ -67,7 +67,7 @@ namespace EntityGraphQL.Schema
         /// <summary>
         /// Given the type TContextType recursively create a query schema based on the public properties of the object. Schema is added into the provider schema
         /// </summary>
-        /// <param name="schema">Schema tp add types to.</param>
+        /// <param name="schema">Schema to add types to.</param>
         /// <param name="autoCreateIdArguments">If true (default), automatically create a field for any root array thats context object contains an Id property. I.e. If Actor has an Id property and the root TContextType contains IEnumerable<Actor> Actors. A root field Actor(id) will be created.</param>
         /// <param name="fieldNamer">Optionally provider a function to generate the GraphQL field name. By default this will make fields names that follow GQL style in lowerCaseCamelStyle</param>
         /// <typeparam name="TContextType"></typeparam>
@@ -79,7 +79,7 @@ namespace EntityGraphQL.Schema
             var contextType = typeof(TContextType);
             var rootFields = GetFieldsFromObject(contextType, schema, autoCreateEnumTypes, autoCreateIdArguments, fieldNamer);
             foreach (var f in rootFields)
-            {             
+            {
                 schema.Query().AddField(f);
             }
             return schema;
@@ -104,7 +104,7 @@ namespace EntityGraphQL.Schema
             // This allows us to "insert" .Select() (and .Include()) before the .First()
             var requiredFieldType = typeof(RequiredField<>).MakeGenericType(idFieldDef.ResolveExpression.Type);
             var fieldNameAndType = new Dictionary<string, Type> { { "id", requiredFieldType } };
-            var argTypes = LinqRuntimeTypeBuilder.GetDynamicType(fieldNameAndType);
+            var argTypes = LinqRuntimeTypeBuilder.GetDynamicType(fieldNameAndType, fieldProp.Name);
             var argTypesValue = Activator.CreateInstance(argTypes);
             var argTypeParam = Expression.Parameter(argTypes, $"args_{argTypes.Name}");
             Type arrayContextType = schemaType.TypeDotnet;
@@ -127,7 +127,7 @@ namespace EntityGraphQL.Schema
                 // If we can't singularize it just use the name plus something as GraphQL doesn't support field overloads
                 name = $"{fieldProp.Name}ById";
             }
-            return new Field(schema, name, selectionExpression, $"Return a {fieldProp.ReturnType.SchemaType.Name} by its Id", argTypesValue, new GqlTypeInfo(fieldProp.ReturnType.SchemaTypeGetter, selectionExpression.Body.Type), fieldProp.RequiredAuthorization);            
+            return new Field(schema, name, selectionExpression, $"Return a {fieldProp.ReturnType.SchemaType.Name} by its Id", argTypesValue, new GqlTypeInfo(fieldProp.ReturnType.SchemaTypeGetter, selectionExpression.Body.Type), fieldProp.RequiredAuthorization);
         }
 
         public static List<Field> GetFieldsFromObject(Type type, ISchemaProvider schema, bool createEnumTypes, bool autoCreateIdArguments, Func<string, string> fieldNamer, bool createNewComplexTypes = true, bool isInputType = false)
@@ -207,7 +207,7 @@ namespace EntityGraphQL.Schema
             {
                 // add non-pural field with argument of ID
                 var idArgField = AddFieldWithIdArgumentIfExists(schema, prop.ReflectedType, field, fieldNamer);
-                if(idArgField != null)
+                if (idArgField != null)
                 {
                     yield return idArgField;
                 }
@@ -250,7 +250,7 @@ namespace EntityGraphQL.Schema
                     if (method == null)
                         throw new Exception($"Could not find {addMethod} method on schema");
                     method = method.MakeGenericMethod(propType);
-                    var typeAdded = (ISchemaType)method.Invoke(schema, new object[] { propType.Name, description })!;                    
+                    var typeAdded = (ISchemaType)method.Invoke(schema, new object[] { propType.Name, description })!;
                     typeAdded.RequiredAuthorization = schema.AuthorizationService.GetRequiredAuthFromType(propType);
 
                     var fields = GetFieldsFromObject(propType, schema, createEnumTypes, autoCreateIdArguments, fieldNamer, createNewComplexTypes, isInputType);